Migrate Trustee Reducer
Create trustee reducer file to maintain with common react principles isolating reducers. Reducer kept at top level due to small size of application
